+++ CHANGELOG.md
@@ -7,10 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### `jsonschema-generator`
 #### Added
+- Convenience constructor on `SchemaGeneratorConfigBuilder` without explicit `ObjectMapper` instance being expected
 - Convenience methods on `FieldScope`/`MethodScope` for accessing (first level) container item annotations: `getContainerItemAnnotation()` and `getContainerItemAnnotationConsideringFieldAndGetter()`
 
 #### Changed
 - `MethodScope.getAnnotation()` now also considers annotations directly on return type (when no matching annotation was found on the method itself)
+
+#### Fixed
+- Attributes set in custom property definitions should not be overridden by other configurations (except for explicit attribute overrides)
 
 ### `jsonschema-module-javax-validation`
 #### Changed